requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
mypy-extensions==1.0.0
mypy==1.3.0
pip==23.1.2
pydantic==1.10.8
PyYAML==6.0
setuptools==60.2.0
types-aiofiles==23.1.0.3
types-babel==2.11.0.14
types-cachetools==5.3.0.5
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.10
types-setuptools==67.8.0.0
typing-extensions==4.6.3
wheel==0.37.1
yarl==1.9.2
